+++ R/step_normalize.R
@@ -11,7 +11,9 @@ orbital.step_normalize <- function(x, all_vars, ...) {
   means <- means[used_vars]
   sds <- sds[used_vars]
 
-  out <- glue::glue("({names(means)} - {means}) / {sds}")
+  out <- glue::glue(
+    "({names(means)} - {format_numeric(means)}) / {format_numeric(sds)}"
+  )
 
   names(out) <- names(means)
   out